requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
spotipy==2.23.0
urllib3==2.1.0
yarl==1.9.4
